# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Carteiras de jogadores e operações de provedores de jogos: apostas, prêmios, derrotas, estornos e reversões.",
        "title": "pedro-test",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/wallets": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria a carteira na moeda do saldo inicial. Com saldo positivo, cria também a transação OPENING, o crédito no ledger e os eventos, no mesmo commit. Com saldo zero, cria só a carteira. Restrita ao serviço interno.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Abre a carteira de um jogador",
                "parameters": [
                    {
                        "description": "Jogador e saldo inicial",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wallet.OpenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/wallet.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    }
                }
            }
        },
        "/wallets/{walletId}": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devolve o saldo e a versão atuais. Restrita ao serviço interno.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Lê uma carteira",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id da carteira",
                        "name": "walletId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wallet.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    }
                }
            }
        },
        "/wallets/{walletId}/ledger": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paginado por cursor opaco, do lançamento mais antigo para o mais novo. A ordem é a posição que o banco deu a cada lançamento, que não muda nem se repete: um lançamento gravado enquanto o cliente lê aparece numa página posterior, e nenhum é pulado nem visto duas vezes. Restrita ao serviço interno.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Lê o ledger de uma carteira",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id da carteira",
                        "name": "walletId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cursor devolvido pela página anterior",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Tamanho da página, de 1 a 200. Padrão 50.",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wallet.LedgerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    }
                }
            }
        },
        "/wallets/{walletId}/reconciliation": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reconstrói o saldo a partir do ledger, abertura incluída, e compara com o saldo gravado, os dois lidos de uma mesma visão consistente dos dados. Não altera nada: uma divergência vai na resposta, no log e numa métrica, e nunca é corrigida. Restrita ao serviço interno.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Reconcilia uma carteira",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id da carteira",
                        "name": "walletId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wallet.ReconciliationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "structs.APIError": {
            "type": "object",
            "properties": {
                "failureCode": {
                    "description": "FailureCode is the stable, documented reason of a business rejection or of an invalid\noperation. It is what a provider reads to tell an input it can correct from a definitive\noutcome, and it is absent on errors that are not about the operation itself.",
                    "type": "string",
                    "example": "INSUFFICIENT_FUNDS"
                },
                "message": {
                    "type": "string",
                    "example": "invalid token"
                }
            }
        },
        "structs.MoneyDTO": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "25.00"
                },
                "currency": {
                    "type": "string",
                    "example": "BRL"
                }
            }
        },
...
        "wallet.LedgerEntryResponse": {
            "type": "object",
            "properties": {
                "balanceAfter": {
                    "$ref": "#/definitions/structs.MoneyDTO"
                },
                "balanceBefore": {
                    "$ref": "#/definitions/structs.MoneyDTO"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2026-09-08T12:00:00.000Z"
                },
                "direction": {
                    "type": "string",
                    "enum": [
                        "DEBIT",
                        "CREDIT"
                    ],
                    "example": "DEBIT"
                },
                "id": {
                    "type": "string",
                    "example": "0192f2a1-4d3e-7b1a-9c55-0f1e2d3c4b5a"
                },
                "money": {
                    "$ref": "#/definitions/structs.MoneyDTO"
                },
                "transactionId": {
                    "type": "string",
                    "example": "0192f298-345e-7e38-af88-e43f851a819d"
                }
            }
        },
        "wallet.LedgerResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wallet.LedgerEntryResponse"
                    }
                },
                "nextCursor": {
                    "description": "NextCursor is absent on the last page. It is opaque: hand it back unchanged.",
                    "type": "string",
                    "example": "djEuNDI"
                }
            }
        },
        "wallet.OpenRequest": {
            "type": "object",
            "properties": {
                "initialBalance": {
                    "$ref": "#/definitions/structs.MoneyDTO"
                },
                "playerId": {
                    "type": "string",
                    "example": "0192f28f-5dc0-7d58-bdb2-814ad6a0f4a1"
                }
            }
        },
        "wallet.ReconciliationResponse": {
            "type": "object",
            "properties": {
                "calculatedBalance": {
                    "$ref": "#/definitions/structs.MoneyDTO"
                },
                "checkedEntries": {
                    "type": "integer",
                    "example": 2
                },
                "consistent": {
                    "type": "boolean",
                    "example": true
                },
                "difference": {
                    "description": "Difference is the stored balance minus the calculated one.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/structs.MoneyDTO"
                        }
                    ]
                },
                "storedBalance": {
                    "$ref": "#/definitions/structs.MoneyDTO"
                },
                "walletId": {
                    "type": "string",
                    "example": "0192f291-27dd-7d3f-8071-5f8685deef37"
                }
            }
        },
        "wallet.Response": {
            "type": "object",
            "properties": {
                "balance": {
                    "$ref": "#/definitions/structs.MoneyDTO"
                },
                "id": {
                    "type": "string",
                    "example": "0192f291-27dd-7d3f-8071-5f8685deef37"
                },
                "playerId": {
                    "type": "string",
                    "example": "0192f28f-5dc0-7d58-bdb2-814ad6a0f4a1"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Cole \"Bearer\" e um token. É o caminho das rotas que exigem um papel de serviço (`internal_service` ou `provider`): o Authorize acima faz password grant, e nenhum usuário do realm tem esses papéis. Pegue o token com `make token-internal` ou `make token-provider`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "http://localhost:8080/realms/pedro-test/protocol/openid-connect/token",
            "scopes": {
                "openid": "Acesso à API"
            }
        }
    }
}